ml-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Pillow==10.1.0
opencv-python-headless==4.8.1.78
